✨ Feat : sns 리다이렉트 url 변경
